Remove dev flag for install
When installing deps in a production environment, the dev dependencies are often not included. However to use this package we need to include the trait in the models which is in our production code. If the dev dependencies aren't included, then this causes an exception saying `Trait "Imanghafoori\EloquentMockery\MockableModel" not found`